+++ ruby/bamboo/block/BlockRiceField.java
@@ -7,7 +7,6 @@ import ruby.bamboo.CustomRenderHandler;
 import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import net.minecraft.block.Block;
-import net.minecraft.block.BlockFarmland;
 import net.minecraft.block.BlockFluid;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
@@ -114,7 +113,7 @@ public class BlockRiceField extends BlockFluid
     @Override
     public AxisAlignedBB getCollisionBoundingBoxFromPool(World par1World, int par2, int par3, int par4)
     {
-        return AxisAlignedBB.getAABBPool().getAABB((double)par2, (double)par3, (double)par4, (double)par2 + 1, (double)par3 + 0.5F, (double)par4 + 1);
+        return AxisAlignedBB.getAABBPool().getAABB(par2, par3, par4, (double)par2 + 1, (double)par3 + 0.5F, (double)par4 + 1);
     }
 
     @Override
