+++ ruby/bamboo/dispenser/BehaviorProjectileDispenseEX.java
@@ -19,7 +19,7 @@ public abstract class BehaviorProjectileDispenseEX extends BehaviorProjectileDis
         IPosition iposition = BlockDispenser.getIPositionFromBlockSource(par1IBlockSource);
         EnumFacing enumfacing = BlockDispenser.getFacing(par1IBlockSource.getBlockMetadata());
         IProjectile iprojectile = this.getProjectileEntity(world, iposition, par2ItemStack.getItemDamage());
-        iprojectile.setThrowableHeading((double)enumfacing.getFrontOffsetX(), (double)((float)enumfacing.getFrontOffsetY() + 0.1F), (double)enumfacing.getFrontOffsetZ(), this.func_82500_b(), this.func_82498_a());
+        iprojectile.setThrowableHeading(enumfacing.getFrontOffsetX(), enumfacing.getFrontOffsetY() + 0.1F, enumfacing.getFrontOffsetZ(), this.func_82500_b(), this.func_82498_a());
         world.spawnEntityInWorld((Entity)iprojectile);
         par2ItemStack.splitStack(1);
         return par2ItemStack;
