+++ ruby/bamboo/entity/EntityKaginawa.java
@@ -1,11 +1,7 @@
 package ruby.bamboo.entity;
 
 import ruby.bamboo.KaginawaHandler;
-import net.minecraft.client.entity.EntityPlayerSP;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.EntityLiving;
 import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.entity.player.EntityPlayerMP;
 import net.minecraft.entity.projectile.EntityThrowable;
 import net.minecraft.util.MathHelper;
 import net.minecraft.util.MovingObjectPosition;
@@ -31,22 +27,22 @@ public class EntityKaginawa extends EntityThrowable
     public void setThrowableHeading(double par1, double par3, double par5, float par7, float par8)
     {
         float f2 = MathHelper.sqrt_double(par1 * par1 + par3 * par3 + par5 * par5);
-        par1 /= (double)f2;
-        par3 /= (double)f2;
-        par5 /= (double)f2;
+        par1 /= f2;
+        par3 /= f2;
+        par5 /= f2;
         par7 *= 2;
-        par1 += 0.5 * 0.007499999832361937D * (double)par8;
-        par3 += 0.5 * 0.007499999832361937D * (double)par8;
-        par5 += 0.5 * 0.007499999832361937D * (double)par8;
-        par1 *= (double)par7;
-        par3 *= (double)par7;
-        par5 *= (double)par7;
+        par1 += 0.5 * 0.007499999832361937D * par8;
+        par3 += 0.5 * 0.007499999832361937D * par8;
+        par5 += 0.5 * 0.007499999832361937D * par8;
+        par1 *= par7;
+        par3 *= par7;
+        par5 *= par7;
         this.motionX = par1;
         this.motionY = par3;
         this.motionZ = par5;
         float f3 = MathHelper.sqrt_double(par1 * par1 + par5 * par5);
         this.prevRotationYaw = this.rotationYaw = (float)(Math.atan2(par1, par5) * 180.0D / Math.PI);
-        this.prevRotationPitch = this.rotationPitch = (float)(Math.atan2(par3, (double)f3) * 180.0D / Math.PI);
+        this.prevRotationPitch = this.rotationPitch = (float)(Math.atan2(par3, f3) * 180.0D / Math.PI);
     }
     @Override
     public void onUpdate()
